V2
Переделал всю архитектуру кода, сделал код более чистым. Переместил все endpoints в одно место, а также все маршруты перенес в ebay.yaml. Сделал для всех сервисов unit-тесты. Создал новую таблицу EbayListing, куда сохраняется вся информация о товаре после его добавления.

+++ src/Service/EbayOffersService.php
@@ -2,6 +2,8 @@
 
 namespace App\Service;
 
+use App\Config\EbayApiEndpoints;
+use App\Entity\EbayListing;
 use App\Entity\InventoryItem;
 use App\Repository\InventoryItemRepository;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
@@ -13,91 +15,80 @@ class EbayOffersService
     private $client;
     private $logger;
     private $tokenService;
-    private $inventoryItemRepository; 
+    private $inventoryItemRepository;
     private string $Environment;
+    private string $baseUrl;
+    private EbayApiRateLimiter $rateLimiter;
 
-    public function __construct(Client $client, TokenService $tokenService, LoggerInterface $logger, InventoryItemRepository $inventoryItemRepository, ParameterBagInterface $params)
-    {
-        $this->client = $client;
+    public function __construct(
+        Client $client = null,
+        TokenService $tokenService,
+        LoggerInterface $logger,
+        InventoryItemRepository $inventoryItemRepository,
+        ParameterBagInterface $params,
+        EbayApiRateLimiter $rateLimiter,
+    ) {
+        $this->client = $client ?? new Client();
         $this->tokenService = $tokenService;
         $this->logger = $logger;
         $this->inventoryItemRepository = $inventoryItemRepository;
-        $this->Environment = $params->get('ebay_environment');
+        $this->rateLimiter = $rateLimiter;
+        $environment = $params->get('ebay_environment');
+        $this->baseUrl = EbayApiEndpoints::getBaseApiUrl($environment);
     }
 
-    private function getApiUrl(string $endpoint): string // Метод для формирования URL в зависимости от среды
-    {
-        $baseUrl = match ($this->Environment) {
-            'Production' => 'https://api.ebay.com',
-            'Sandbox' => 'https://api.sandbox.ebay.com',
-            default => throw new \InvalidArgumentException('Invalid environment specified.'),
-        };
-
-        return $baseUrl . $endpoint;
-    }
-    
-    public function createOfferFromDatabaseItem(string $sku): ?array // Метод для создания предложения товара на eBay на основе данных из базы по SKU
+    /**
+     * Метод для создания предложения товара на eBay на основе данных из базы по SKU
+     *
+     * @param string $sku SKU товара.
+     * @return array|null Ответ от API eBay после создания предложения или null в случае ошибки.
+     */
+    public function createOfferFromDatabaseItem(string $sku): ?array 
     {
         try {
-            $item = $this->inventoryItemRepository->findBySku($sku); // Получаем товар из базы данных по SKU
+            // Получаем товар из базы данных по SKU
+            $item = $this->inventoryItemRepository->findBySku($sku); 
 
-            if (!$item) { // Если товар не найден, логируем предупреждение и возвращаем null
+            if (!$item) { 
                 $this->logger->warning('Item not found in database with SKU: ' . $sku);
                 return null;
             }
 
-            $listingDescription = $item->getDescription();
-            $listingDescriptionHtml = "<p>" . $listingDescription . "</p>";
-            
-            $offerData = [
-                'sku' => $item->getSku(),
-                'marketplaceId' => $item->getMarketplaceId(),
-                'format' => $item->getFormat(), 
-                'listingDescription' => $listingDescriptionHtml,
-                'availableQuantity' => $item->getAvailabilityQuantity(),
-                'pricingSummary' => [
-                    'price' => [
-                        'value' => $item->getPrice(),
-                        'currency' => $item->getCurrency(),
-                    ],
-                ],
-                'merchantLocationKey' => 'TEST1', // надо указать локацию 
-                'categoryId' => '58058',
-                'listingPolicies' => [
-                    'fulfillmentPolicyId' => 'TEST1', // так же надо указать политики
-                    'paymentPolicyId' => 'TEST1',
-                    'returnPolicyId' => 'TEST1',
-                ],
-            ];
+            // Получаем ACCESS TOKEN из метода
+            $accessToken = $this->getAccessToken();
 
-            $accessToken = $this->tokenService->getValidAccessToken(); // Получаем токен доступа для eBay
-            if ($accessToken) {
-                $this->logger->info('Access token successfully retrieved for SKU: ' . $item->getSku());
-            } else {
-                $this->logger->warning('No access token retrieved for SKU: ' . $item->getSku());
+            if (!$accessToken) {
+                $this->logger->warning("No access token retrieved");
+                return null;
             }
 
-            $url = $this->getApiUrl('/sell/inventory/v1/offer'); // Формируем URL для создания предложения на eBay
+            // проверка лимита запросов
+            if (!$this->rateLimiter->incrementAndCheck()) {
+                $this->logger->warning("eBay API call not executed: daily limit has been exceeded.");
+                return null;
+            }
 
-            $response = $this->client->post($url, [ // Отправляем POST запрос для создания предложения товара
-                'headers' => [
-                    'Authorization' => 'Bearer ' . $accessToken,
-                    'Accept' => 'application/json',
-                    'Content-Type' => 'application/json',
-                    'Content-Language' => 'en-US',
-                ],
+            // Формируем URL для запроса создания offer
+            $url = $this->baseUrl . '/sell/inventory/v1/offer'; 
+            // получаем body пердмета из метода
+            $offerData = $this->buildOfferData($item);
+
+            // отправляем запрос на создания offer
+            $response = $this->client->post($url, [ 
+                'headers' => $this->buildJsonHeaders($accessToken),
                 'json' => $offerData,
             ]);
 
-            $responseBody = $response->getBody()->getContents(); // Получаем содержимое ответа от eBay
+            $responseBody = $response->getBody()->getContents(); 
             $this->logger->info('Response from eBay: ' . $responseBody);
 
-            if ($response->getStatusCode() === 200) { // Проверяем код ответа
-                return json_decode($responseBody, true); // Возвращаем данные предложения
+            if ($response->getStatusCode() === 200) {
+                $responseData = json_decode($responseBody, true);
+                return $responseData;
             }
 
-            $responseData = json_decode($responseBody, true); // Если ответ содержит ошибки, возвращаем их
-            return $responseData ? $responseData : null;
+            $responseData = json_decode($responseBody, true); 
+            return $responseData ?: null;
 
         } catch (\Exception $e) {
             $this->logger->error('Error creating offer for SKU: ' . $sku . ' - ' . $e->getMessage());
@@ -105,50 +96,78 @@ class EbayOffersService
         }
     }
 
-    public function getOffers(string $sku): ?array // Метод для получения предложения по SKU с eBay
+    /**
+     * Метод для получения предложения по SKU с eBay
+     *
+     * @param string $sku SKU товара.
+     * @return array|null Ответ от API eBay со списком предложений или null в случае ошибки.
+     */
+    public function getOffers(string $sku): ?array
     {
         try {
-            $accessToken = $this->tokenService->getValidAccessToken(); // Получаем токен доступа для eBay
-            if ($accessToken) {
-                $this->logger->info('Access token successfully retrieved');
-            } else {
-                $this->logger->warning('No access token retrieved');
+            // Получаем Token из метода
+            $accessToken = $this->getAccessToken();
+
+            // Если Token не найден, логируем предупреждение и возвращаем null
+            if (!$accessToken) {
+                $this->logger->warning("No access token retrieved");
+                return null;
             }
-            
-            $url = $this->getApiUrl('/sell/inventory/v1/offer?sku=' . $sku);// Формируем URL для запроса предложения по SKU
-    
-            $response = $this->client->get($url, [ // Отправляем GET запрос для получения предложения по SKU
-                'headers' => [
-                    'Authorization' => 'Bearer ' . $accessToken,
-                    'Accept' => 'application/json',
-                ],
+
+            // проверка лимита запросов
+            if (!$this->rateLimiter->incrementAndCheck()) {
+                $this->logger->warning("eBay API call not executed: daily limit has been exceeded.");
+                return null;
+            }
+
+            // Формируем URL для запроса предложения по SKU
+            $url =  $this->baseUrl . '/sell/inventory/v1/offer?sku=' . $sku;
+
+            // Отправляем GET запрос для получения предложения по SKU
+            $response = $this->client->get($url, [ 
+                'headers' => $this->buildAuthHeaders($accessToken),
             ]);
     
             $responseBody = $response->getBody()->getContents();
-            return json_decode($responseBody, true); // Декодируем и возвращаем данные из ответа
+
+            // Декодируем и возвращаем данные из ответа
+            return json_decode($responseBody, true); 
         } catch (\Exception $e) {
             $this->logger->error('Error fetching offer by SKU - ' . $e->getMessage());
             return null;
         }
     }
 
-    public function getOfferByOfferId(string $offerId): ?array // Метод для получения предложения по offerId с eBay
+    /**
+     * Метод для получения предложения по offerId с eBay
+     *
+     * @param string $offerId ID предложения.
+     * @return array|null Ответ от API eBay с информацией о предложении или null в случае ошибки.
+     */
+    public function getOfferByOfferId(string $offerId): ?array
     {
         try {
-            $accessToken = $this->tokenService->getValidAccessToken(); // Получаем токен доступа для eBay
-            if ($accessToken) {
-                $this->logger->info('Access token successfully retrieved');
-            } else {
-                $this->logger->warning('No access token retrieved');
+            // Получаем Token из метода
+            $accessToken = $this->getAccessToken();
+
+            // Если Token не найден, логируем предупреждение и возвращаем null
+            if (!$accessToken) {
+                $this->logger->warning("No access token retrieved");
+                return null;
+            }
+
+            // проверка лимита запросов
+            if (!$this->rateLimiter->incrementAndCheck()) {
+                $this->logger->warning("eBay API call not executed: daily limit has been exceeded.");
+                return null;
             }
     
-            $url = $this->getApiUrl('/sell/inventory/v1/offer/' . $offerId);// Формируем URL для запроса предложения по offerId
+            // Формируем URL для запроса предложения по offerId
+            $url = $this->baseUrl . '/sell/inventory/v1/offer/' . $offerId;
 
-            $response = $this->client->get($url, [ // Отправляем GET запрос для получения предложения по offerId
-                'headers' => [
-                    'Authorization' => 'Bearer ' . $accessToken,
-                    'Accept' => 'application/json',
-                ],
+            // Отправляем GET запрос для получения предложения по offerId
+            $response = $this->client->get($url, [ 
+                'headers' => $this->buildAuthHeaders($accessToken),
             ]);
     
             $responseBody = $response->getBody()->getContents();
@@ -159,22 +178,36 @@ class EbayOffersService
         }
     }
 
-    public function deleteOfferByOfferId(string $offerId): bool // Метод для удаления предложения по offerId с eBay
+    /**
+     * Метод для получения предложения по offerId с eBay
+     *
+     * @param string $offerId ID предложения для удаления.
+     * @return bool True в случае успешного удаления, false в случае ошибки.
+     */
+    public function deleteOfferByOfferId(string $offerId): bool
     {
         try {
-            $accessToken = $this->tokenService->getValidAccessToken(); // Получаем токен доступа для eBay
+            // Получаем Token из метода
+            $accessToken = $this->getAccessToken();
+
+            // Если Token не найден, логируем предупреждение и возвращаем null
             if (!$accessToken) {
-                $this->logger->warning("Access token not found for offerId: $offerId");
-                return false;
+                $this->logger->warning("No access token retrieved");
+                return null;
+            }
+
+            // проверка лимита запросов
+            if (!$this->rateLimiter->incrementAndCheck()) {
+                $this->logger->warning("eBay API call not executed: daily limit has been exceeded.");
+                return null;
             }
     
-            $url = $this->getApiUrl('/sell/inventory/v1/offer/' . $offerId); // Формируем URL для удаления предложения по offerId
+            // Формируем URL для удаления предложения по offerId
+            $url =  $this->baseUrl . '/sell/inventory/v1/offer/' . $offerId; 
     
+            // Отправляем DELET запрос для удаления предложения по offerId
             $response = $this->client->delete($url, [
-                'headers' => [
-                    'Authorization' => 'Bearer ' . $accessToken,
-                    'Accept' => 'application/json',
-                ],
+                'headers' => $this->buildAuthHeaders($accessToken),
             ]);
     
             if ($response->getStatusCode() === 204) {
@@ -190,34 +223,133 @@ class EbayOffersService
         }
     }
 
-    public function publishOffer(string $offerId): ?array // Метод для публикации оффера по offerId
+    /**
+     * Метод для публикации оффера по offerId
+     *
+     * @param string $offerId ID предложения для публикации.
+     * @return array|null Ответ от API eBay после публикации предложения или null в случае ошибки.
+     */
+    public function publishOffer(string $offerId): ?array 
     {
         try {
-            $accessToken = $this->tokenService->getValidAccessToken(); // Получаем токен доступа для eBay
+            // Получаем Token из метода
+            $accessToken = $this->getAccessToken();
 
-            if ($accessToken) {
-                $this->logger->info('Access token successfully retrieved for publishing offer');
-            } else {
-                $this->logger->warning('No access token retrieved for publishing offer');
+            // Если Token не найден, логируем предупреждение и возвращаем null
+            if (!$accessToken) {
+                $this->logger->warning("No access token retrieved");
                 return null;
             }
 
-            $url = $this->getApiUrl('/sell/inventory/v1/offer/' . $offerId . '/publish'); // Формируем URL публикации
+            // проверка лимита запросов
+            if (!$this->rateLimiter->incrementAndCheck()) {
+                $this->logger->warning("eBay API call not executed: daily limit has been exceeded.");
+                return null;
+            }
 
-            $response = $this->client->post($url, [ // POST-запрос на публикацию оффера
-                'headers' => [
-                    'Authorization' => 'Bearer ' . $accessToken,
-                    'Accept'        => 'application/json',
-                    'Content-Type'  => 'application/json',
-                ],
+            // Формируем URL публикации
+            $url =  $this->baseUrl . '/sell/inventory/v1/offer/' . $offerId . '/publish'; 
+
+            // POST-запрос на публикацию оффера
+            $response = $this->client->post($url, [ 
+                'headers' => $this->buildJsonHeaders($accessToken),
             ]);
 
             $responseBody = $response->getBody()->getContents();
-            return json_decode($responseBody, true); // Возвращаем ответ от eBay
+
+            // Возвращаем ответ от eBay
+            return json_decode($responseBody, true); 
         } catch (\Exception $e) {
             $this->logger->error('Error publishing offer - ' . $e->getMessage());
             return null;
         }
+    }
+
+    /**
+     * Метод для получения AccesToken
+     *
+     * @return string|null Access Token или null, если получить не удалось.
+     */
+    private function getAccessToken(): ?string
+    {
+        $accessToken = $this->tokenService->getValidAccessToken();
+
+        if (!$accessToken) {
+            $this->logger->warning("No access token retrieved");
+        }
+
+        return $accessToken;
+    }
+
+    /**
+     * Метод для для создания Auth Header
+     *
+     * @param string $accessToken Действительный Access Token.
+     * @return array Массив с заголовком 'Authorization'.
+     */
+    private function buildAuthHeaders(string $accessToken): array
+    {
+        return [
+            'Authorization' => "Bearer {$accessToken}",
+            'Accept' => 'application/json',
+        ];
+    }
+
+    /**
+     * Метод для для создания Build Header
+     *
+     * @param string $accessToken Действительный Access Token.
+     * @return array Массив с заголовками 'Authorization', 'Accept' и 'Content-Type'.
+     */
+    private function buildJsonHeaders(string $accessToken): array
+    {
+        return[
+            'Authorization' => 'Bearer ' . $accessToken,
+            'Accept' => 'application/json',
+            'Content-Type' => 'application/json',
+            'Content-Language' => 'en-US',
+        ];
+    }
+
+    /**
+     * Метод для для создания Offer body
+     *
+     * @param InventoryItem $item Сущность товара из базы данных.
+     * @return array Массив с данными предложения в формате, ожидаемом API eBay.
+     */
+    private function buildOfferData(InventoryItem $item): array
+    {
+        return [
+            'sku' => $item->getSku(),
+            'marketplaceId' => $item->getMarketplaceId(),
+            'format' => $item->getFormat(),
+            'listingDescription' => '<p>' . $item->getDescription() . '</p>',
+            'availableQuantity' => $item->getAvailabilityQuantity(),
+            'pricingSummary' => [
+                'price' => [
+                    'value' => $item->getPrice(),
+                    'currency' => $item->getCurrency(),
+                ],
+            ],
+            'merchantLocationKey' => 'TEST1', //Полу которое нужно указать для работы публикации
+            'categoryId' => '58058',
+            "listingPolicies" => [
+                "fulfillmentPolicyId" => "TEST1", //Полу которое нужно указать для работы публикации
+                "paymentPolicyId"=> "TEST1", //Полу которое нужно указать для работы публикации
+                "returnPolicyId" => "TEST1" //Полу которое нужно указать для работы публикации
+            ],
+        ];
+    }
+
+    /**
+     * Получает InventoryItem по SKU.
+     *
+     * @param string $sku SKU товара для поиска.
+     * @return InventoryItem|null Найденный товар или null, если товар не найден.
+     */
+    public function getInventoryItemBySku(string $sku): ?InventoryItem
+    {
+        return $this->inventoryItemRepository->findBySku($sku);
     }
 
 }
